refact(map): rename main map file
